*Player(NEW) added to replace the old one
This player makes use of the command pattern for its actions.
Some of the behaviour classes have been tweaked to use the localscale of the gameobject instead of a charactercontroller2d member. This way it works for all gameobjects.

The player largely is animated, but the jump animation is not factored in yet, I am trying to keep the old animators entirely intact.